Move diagnostic panel, header, and report list view to ui module
