+++ taxel-gui/src/app/report_list.rs
@@ -3,123 +3,23 @@ use crate::{
     report_store::{self, ReportManifestEntry, ReportStatus, ReportStore, ReportSummary},
 };
 use anyhow::Result;
-use eframe::egui::{Color32, CursorIcon, Frame, Label, Margin, RichText, Sense, Ui};
 use std::{
     collections::{HashMap, HashSet},
-    path::{Path, PathBuf},
+    path::Path,
     time::SystemTime,
 };
-
-/// Draws the report list view, showing imported reports and allowing the user
-/// to select one to open.
-pub fn draw_report_list(ui: &mut Ui, reports: &[ReportSummary], loading: bool) -> Option<PathBuf> {
-    let list_width = ui.available_width().min(560.0);
-    let created_col_width = 120.0;
-    let status_col_width = 90.0;
-    let column_spacing = ui.spacing().item_spacing.x * 2.0;
-    let report_col_width =
-        (list_width - created_col_width - status_col_width - column_spacing).max(240.0);
-
-    ui.vertical_centered(|ui| {
-        ui.heading("Your Reports");
-        ui.add_space(6.0);
-    });
-
-    if reports.is_empty() {
-        ui.vertical_centered(|ui| {
-            ui.label("No imported reports yet. Use Import report to add an XML file.");
-        });
-        return None;
-    }
-
-    ui.vertical_centered(|ui| {
-        ui.label(format!("{} report(s)", reports.len()));
-    });
-    ui.add_space(4.0);
-
-    let mut selected = None;
-
-    ui.vertical_centered(|ui| {
-        ui.set_max_width(list_width);
-        ui.set_width(list_width);
-        ui.horizontal(|ui| {
-            ui.add_sized(
-                [created_col_width, 18.0],
-                Label::new(RichText::new("Created")),
-            );
-            ui.add_sized(
-                [report_col_width, 18.0],
-                Label::new(RichText::new("Report")),
-            );
-            ui.add_sized(
-                [status_col_width, 18.0],
-                Label::new(RichText::new("Status")),
-            );
-        });
-        ui.separator();
-
-        for (idx, report) in reports.iter().enumerate() {
-            let row = Frame::new()
-                .fill(Color32::TRANSPARENT)
-                .inner_margin(Margin::symmetric(0, 6))
-                .show(ui, |ui| {
-                    ui.set_width(list_width);
-                    ui.horizontal(|ui| {
-                        ui.add_sized([created_col_width, 18.0], Label::new(&report.created_date));
-                        ui.add_sized(
-                            [report_col_width, 18.0],
-                            Label::new(&report.display_name).truncate(),
-                        );
-                        ui.add_sized(
-                            [status_col_width, 18.0],
-                            Label::new(report.report_status.as_str()),
-                        );
-                    });
-                });
-
-            let response = ui.interact(
-                row.response.rect,
-                ui.id().with(("report_row", idx)),
-                if loading {
-                    Sense::hover()
-                } else {
-                    Sense::click()
-                },
-            );
-            let response = if loading {
-                response
-            } else {
-                response.on_hover_cursor(CursorIcon::PointingHand)
-            };
-
-            if response.hovered() && !loading {
-                ui.painter().rect_filled(
-                    row.response.rect,
-                    2.0,
-                    ui.visuals().widgets.hovered.bg_fill.gamma_multiply(0.25),
-                );
-            }
-
-            if response.clicked() {
-                selected = Some(report.path.clone());
-            }
-        }
-    });
-
-    selected
-}
 
 /// The `ReportList` struct manages the list of imported reports, including
 /// their metadata and creation dates. Creation timestamps are persisted in a
 /// JSON manifest in the reports directory.
-pub(super) struct ReportList {
+pub struct ReportList {
     reports: Vec<ReportSummary>,
     creation_dates: HashMap<String, i64>,
     report_statuses: HashMap<String, ReportStatus>,
 }
 
 impl ReportList {
-    pub(super) fn new() -> Self {
+    pub fn new() -> Self {
         Self {
             reports: Vec::new(),
             creation_dates: HashMap::new(),
@@ -127,7 +27,7 @@ impl ReportList {
         }
     }
 
-    pub(super) fn refresh(&mut self) -> Result<()> {
+    pub fn refresh(&mut self) -> Result<()> {
         let mut reports = ReportStore::load_reports()?;
 
         let manifest = report_store::load_report_manifest()?;
@@ -153,7 +53,7 @@ impl ReportList {
     /// `creation_dates` HashMap. This ensures that the creation date is
     /// preserved even if the report list is refreshed from the filesystem,
     /// which may not retain the original creation date metadata.
-    pub(super) fn register_imported_report(&mut self, report_path: &Path) {
+    pub fn register_imported_report(&mut self, report_path: &Path) {
         let unix_seconds = report_store::system_time_to_unix_seconds(SystemTime::now());
         let report_path = report_path.to_string_lossy().to_string();
         self.creation_dates
@@ -162,12 +62,12 @@ impl ReportList {
             .insert(report_path, ReportStatus::Draft);
     }
 
-    pub(super) fn report_status(&self, report_path: &Path) -> Option<ReportStatus> {
+    pub fn report_status(&self, report_path: &Path) -> Option<ReportStatus> {
         let report_path = report_path.to_string_lossy().to_string();
         self.report_statuses.get(&report_path).copied()
     }
 
-    pub(super) fn set_report_status(
+    pub fn set_report_status(
         &mut self,
         report_path: &Path,
         status: ReportStatus,
@@ -192,7 +92,7 @@ impl ReportList {
         }
     }
 
-    pub(super) fn reports(&self) -> &[ReportSummary] {
+    pub fn reports(&self) -> &[ReportSummary] {
         &self.reports
     }
