+++ taxel-gui/src/ui/mod.rs
@@ -1,2 +1,5 @@
+pub mod diagnostic_panel;
+pub mod header;
+pub mod report_list_view;
 pub mod report_view;
 pub mod widgets;
